+++ doc/examples/effects/sequence_effect/lib/main.dart
@@ -1,3 +1,4 @@
+import 'package:flame/effects/combined_effect.dart';
 import 'package:flame/effects/move_effect.dart';
 import 'package:flame/effects/scale_effect.dart';
 import 'package:flame/effects/rotate_effect.dart';
@@ -34,7 +35,7 @@ class MyGame extends BaseGame with TapDetector {
     greenSquare.clearEffects();
 
     final move1 = MoveEffect(
-      destination: Vector2(dx, dy),
+      path: [Vector2(dx, dy)],
       speed: 250.0,
       curve: Curves.bounceInOut,
       isInfinite: false,
@@ -42,16 +43,20 @@ class MyGame extends BaseGame with TapDetector {
     );
 
     final move2 = MoveEffect(
-      destination: Vector2(dx, dy + 150),
+      path: [
+        Vector2(dx, dy + 50),
+        Vector2(dx - 50, dy - 50),
+        Vector2(dx + 50, dy),
+      ],
       speed: 150.0,
       curve: Curves.easeIn,
       isInfinite: false,
-      isAlternating: true,
+      isAlternating: false,
     );
 
     final scale = ScaleEffect(
       size: Vector2(dx, dy),
-      speed: 250.0,
+      speed: 100.0,
       curve: Curves.easeInCubic,
       isInfinite: false,
       isAlternating: false,
@@ -59,15 +64,21 @@ class MyGame extends BaseGame with TapDetector {
 
     final rotate = RotateEffect(
       radians: (dx + dy) % pi,
-      speed: 2.0,
+      speed: 0.8,
       curve: Curves.decelerate,
       isInfinite: false,
       isAlternating: false,
     );
 
+    final combination = CombinedEffect(
+      effects: [move2, rotate],
+      isAlternating: false,
+      isInfinite: false,
+    );
+
     final sequence = SequenceEffect(
-      effects: [move1, scale, move2, rotate],
-      isInfinite: true,
+      effects: [move1, scale, combination],
+      isInfinite: false,
       isAlternating: true,
     );
     greenSquare.addEffect(sequence);