+++ doc/examples/particles/lib/main.dart
@@ -587,6 +587,7 @@ class TrafficLightComponent extends Component {
 
   @override
   void update(double dt) {
+    super.update(dt);
     colorChangeTimer.update(dt);
   }