+++ lib/effects/combined_effect.dart
@@ -16,6 +16,10 @@ class CombinedEffect extends PositionComponentEffect {
     bool isAlternating = false,
     Function onComplete,
   }) : super(isInfinite, isAlternating, onComplete: onComplete) {
+    assert(
+      effects.every((effect) => effect.component == null),
+      'Each effect can only be added once',
+    );
     final types = effects.map((e) => e.runtimeType);
     assert(
       types.toSet().length == types.length,
@@ -47,7 +51,9 @@ class CombinedEffect extends PositionComponentEffect {
     super.update(dt);
     effects.forEach((effect) => _updateEffect(effect, dt));
     if (effects.every((effect) => effect.hasFinished())) {
-      if (isInfinite) {
+      if (isAlternating && curveDirection.isNegative) {
+        effects.forEach((effect) => effect.isAlternating = true);
+      } else if (isInfinite) {
         reset();
       } else if (isAlternating && isMin()) {
         dispose();
@@ -58,8 +64,13 @@ class CombinedEffect extends PositionComponentEffect {
   @override
   void reset() {
     super.reset();
+    if (component != null) {
+      component.position = originalPosition;
+      component.angle = originalAngle;
+      component.size = originalSize;
+      initialize(component);
+    }
     effects.forEach((effect) => effect.reset());
-    initialize(component);
   }
 
   @override
@@ -83,6 +94,11 @@ class CombinedEffect extends PositionComponentEffect {
     if (isMax()) {
       _maybeReverse(effect);
     }
+  }
+
+  @override
+  bool hasFinished() {
+    return super.hasFinished() && effects.every((e) => e.hasFinished());
   }
 
   void _maybeReverse(PositionComponentEffect effect) {