+++ CHANGELOG.md
@@ -12,6 +12,10 @@
  - Removing all deprecated methods
  - Rename `resize` method on components to `onGameResize`
  - Make `Resizable` have a `gameSize` property instead of `size`
+ - Fix bug with CombinedEffect inside SequenceEffect
+ - Fix wrong end angle for relative rotational effects
+ - Use a list of Vector2 for Move effect to open up for more advanced move effects
+ - Generalize effects api to include all components
 
 ## [next]
  - Fix spriteAsWidget deprecation message